+++ src/hexen/po_man.h
@@ -56,9 +56,9 @@ typedef struct
 void T_PolyDoor(polydoor_t *pd);
 void T_RotatePoly(polyevent_t *pe);
 void T_MovePoly(polyevent_t *pe);
-dbool EV_RotatePoly(line_t *line, byte *args, int direction, dbool overRide);
-dbool EV_MovePoly(line_t *line, byte *args, dbool timesEight, dbool overRide);
-dbool EV_OpenPolyDoor(line_t *line, byte *args, podoortype_t type);
+dbool EV_RotatePoly(line_t *line, int *args, int direction, dbool overRide);
+dbool EV_MovePoly(line_t *line, int *args, dbool timesEight, dbool overRide);
+dbool EV_OpenPolyDoor(line_t *line, int *args, podoortype_t type);
 
 dbool PO_MovePolyobj(int num, int x, int y);
 dbool PO_RotatePolyobj(int num, angle_t angle);
